+++ src/reader/core/search.py
@@ -101,7 +101,15 @@ class Search:
         self.storage = storage
 
     @wrap_storage_exceptions(SearchError)
-    def enable(self) -> None:
+    def enable(self):
+        try:
+            self._enable()
+        except sqlite3.OperationalError as e:
+            if "table entries_search already exists" in str(e).lower():
+                pass
+            return
+
+    def _enable(self) -> None:
         with ddl_transaction(self.storage.db) as db:
 
             # The column names matter, as they can be used in column filters;
@@ -112,7 +120,7 @@ class Search:
             #
             db.execute(
                 """
-                CREATE VIRTUAL TABLE IF NOT EXISTS entries_search USING fts5(
+                CREATE VIRTUAL TABLE entries_search USING fts5(
                     title,  -- entries.title
                     content,  -- entries.summary or one of entries.content
                     feed,  -- feeds.title or feed.user_title
@@ -134,7 +142,7 @@ class Search:
 
             db.execute(
                 """
-                CREATE TABLE IF NOT EXISTS entries_search_sync_state (
+                CREATE TABLE entries_search_sync_state (
                     id TEXT NOT NULL,
                     feed TEXT NOT NULL,
                     to_update INTEGER NOT NULL DEFAULT 1,
@@ -157,7 +165,7 @@ class Search:
 
             db.execute(
                 """
-                CREATE TRIGGER IF NOT EXISTS entries_search_entries_insert
+                CREATE TRIGGER entries_search_entries_insert
                 AFTER INSERT ON entries
                 BEGIN
                     INSERT INTO entries_search_sync_state
@@ -167,7 +175,7 @@ class Search:
             )
             db.execute(
                 """
-                CREATE TRIGGER IF NOT EXISTS entries_search_entries_update
+                CREATE TRIGGER entries_search_entries_update
                 AFTER UPDATE ON entries
                 BEGIN
                     UPDATE entries_search_sync_state
@@ -181,7 +189,7 @@ class Search:
             )
             db.execute(
                 """
-                CREATE TRIGGER IF NOT EXISTS entries_search_entries_delete
+                CREATE TRIGGER entries_search_entries_delete
                 AFTER DELETE ON entries
                 BEGIN
                     UPDATE entries_search_sync_state
@@ -199,7 +207,7 @@ class Search:
             # the entries delete trigger will take care of its entries.
             db.execute(
                 """
-                CREATE TRIGGER IF NOT EXISTS entries_search_feeds_update
+                CREATE TRIGGER entries_search_feeds_update
                 AFTER UPDATE ON feeds
                 BEGIN
                     UPDATE entries_search_sync_state